# backend/api-doc.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/v1/category/update-category/{id}": {
      "put": {
        "tags": [
          "category-controller"
        ],
        "summary": "Update category",
        "operationId": "updateCategory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseSuccessObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/author/update-author/{id}": {
      "put": {
        "tags": [
          "author-controller"
        ],
        "summary": "Update author",
        "operationId": "updateAuthor",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseSuccessObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/category/add-category": {
      "post": {
        "tags": [
          "category-controller"
        ],
        "summary": "Add category",
        "operationId": "addCategory",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseSuccessObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/author/add-author": {
      "post": {
        "tags": [
          "author-controller"
        ],
        "summary": "Add author",
        "operationId": "addAuthor",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseSuccessObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/category/get-category-by-id/{id}": {
      "get": {
        "tags": [
          "category-controller"
        ],
        "summary": "Get category by id",
        "operationId": "getCategoryById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseSuccessObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/author/get-author-by-id/{id}": {
      "get": {
        "tags": [
          "author-controller"
        ],
        "operationId": "getAuthorById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseSuccessObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/category/delete-category/{id}": {
      "delete": {
        "tags": [
          "category-controller"
        ],
        "summary": "Delete category",
        "operationId": "deleteCategory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseSuccessObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/author/delete-author/{id}": {
      "delete": {
        "tags": [
          "author-controller"
        ],
        "summary": "Delete author",
        "operationId": "deleteAuthor",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseSuccessObject"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ResponseSuccessObject": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object"
          }
        }
      },
...
      "CategoryRequest": {
        "required": [
          "name",
          "userId"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "parentId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "AuthorRequest": {
        "required": [
          "name",
          "userId"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
...
    }
  }
}